test: migrate Alert, Confirm tests
