+++ src/components/Dialogs/Confirm/__tests__/Confirm.spec.tsx
@@ -2,42 +2,45 @@ import React from 'react';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import '@testing-library/jest-dom';
-import { Alert } from '../Alert';
+import { Confirm } from '../Confirm';
 
-describe('Alert', () => {
+describe('Confirm', () => {
     const mockOnAccept = jest.fn();
+    const mockOnCancel = jest.fn();
     let view;
 
     beforeEach(() => {
         view = render(
-            <Alert
+            <Confirm
                 isOpen
                 acceptButton={{ onClick: mockOnAccept, label: 'OK' }}
+                cancelButton={{ onClick: mockOnCancel, label: 'Cancel' }}
                 ariaHideApp={false}
-                contentLabel="Content label"
+                contentLabel="Content Label"
             >
-                Body of the alert
-            </Alert>
+                Body of the confirm
+            </Confirm>
         );
     });
 
     it('should render correctly', () => {
         expect(view.container).toMatchSnapshot();
         expect(screen.queryAllByRole('heading')).toHaveLength(0);
-        expect(screen.getAllByRole('button')).toHaveLength(1);
+        expect(screen.getAllByRole('button')).toHaveLength(2);
     });
 
     it('should render correctly with title', () => {
         view.rerender(
-            <Alert
+            <Confirm
                 isOpen
                 acceptButton={{ onClick: mockOnAccept, label: 'OK' }}
+                cancelButton={{ onClick: mockOnCancel, label: 'Cancel' }}
                 ariaHideApp={false}
-                contentLabel="Content label"
+                contentLabel="Content Label"
                 title="Title"
             >
-                Body of the alert
-            </Alert>
+                Body of the confirm
+            </Confirm>
         );
 
         expect(view.container).toMatchSnapshot();
@@ -46,8 +49,14 @@ describe('Alert', () => {
 
     it('should call onAccept cb when button is clicked', async () => {
         const user = userEvent.setup();
-        await user.click(screen.getByRole('button'));
+        await user.click(screen.getAllByRole('button')[1]);
 
         expect(mockOnAccept).toHaveBeenCalledTimes(1);
+    });
+    it('should call onAccept cb when button is clicked', async () => {
+        const user = userEvent.setup();
+        await user.click(screen.getAllByRole('button')[0]);
+
+        expect(mockOnCancel).toHaveBeenCalledTimes(1);
     });
 });
